Project/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.1.12
djangorestframework==3.14.0
djongo==1.3.6
idna==3.4
pymongo==3.12.3
pytz==2022.4
sqlparse==0.2.4
typing-extensions==4.4.0
urllib3==1.26.12
